+++ src/main/java/com/phantomwing/rusticdelight/datagen/ModItemModelProvider.java
@@ -39,8 +39,10 @@ public class ModItemModelProvider extends ItemModelProvider {
         simpleItem(ModItems.CHERRY_BLOSSOM_CHEESECAKE);
         simpleItem(ModItems.CHERRY_BLOSSOM_CHEESECAKE_SLICE);
         simpleItem(ModItems.CHERRY_BLOSSOM_COOKIE);
+        simpleItem(ModItems.CHERRY_BLOSSOM_PANCAKE);
         simpleItem(ModItems.CHERRY_BLOSSOM_PANCAKES);
         simpleItem(ModItems.CHERRY_BLOSSOM_ROLL);
+        simpleItem(ModItems.CHOCOLATE_PANCAKE);
         simpleItem(ModItems.CHOCOLATE_PANCAKES);
         simpleItem(ModItems.COFFEE_BEANS);
         simpleItem(ModItems.COFFEE_BRAISED_BEEF);
@@ -54,10 +56,12 @@ public class ModItemModelProvider extends ItemModelProvider {
         simpleItem(ModItems.FRIED_MUSHROOMS);
         simpleItem(ModItems.FRUIT_BEIGNET);
         simpleItem(ModItems.GOLDEN_COFFEE_BEANS);
+        simpleItem(ModItems.HONEY_PANCAKE);
         simpleItem(ModItems.HONEY_PANCAKES);
         simpleItem(ModItems.MILK_COFFEE);
         simpleItem(ModItems.POTATO_SALAD);
         simpleItem(ModItems.POTATO_SLICES);
+        simpleItem(ModItems.PUMPKIN_PANCAKE);
         simpleItem(ModItems.PUMPKIN_PANCAKES);
         simpleItem(ModItems.ROASTED_BELL_PEPPER_GREEN);
         simpleItem(ModItems.ROASTED_BELL_PEPPER_YELLOW);
@@ -69,6 +73,7 @@ public class ModItemModelProvider extends ItemModelProvider {
         simpleItem(ModItems.STUFFED_BELL_PEPPER_YELLOW);
         simpleItem(ModItems.STUFFED_BELL_PEPPER_RED);
         simpleItem(ModItems.SPRING_ROLLS);
+        simpleItem(ModItems.VEGETABLE_PANCAKE);
         simpleItem(ModItems.VEGETABLE_PANCAKES);
 
         // Blocks